pyproject.toml
[tool.poetry]
name = "olime-backend-v2"
version = "0.1.0"
description = ""
authors = ["naturalthwn <naturalthwns@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "olime_backend_v2"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.24.2"
Pillow = "^9.3.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
